supply-control/requirements.txt
Flask
Flask-SQLAlchemy
Flask-WTF
WTForms
Flask-Login
reportlab
requests
python-barcode